requirements.txt
discord.py
requests
PythonGists
bs4
lxml
Pillow
mtranslate
pyfiglet
